Feat: Refactor on Fetching Curent TBILLS

+++ Quotations Board Backend/DTOs/TBill/HistoricalTbill.cs
@@ -4,7 +4,7 @@ public class HistoricalTbill
     // public string IssueNumber { get; set; } = null!;
     public DateTime IssueDate { get; set; }
     public DateTime MaturityDate { get; set; }
-    public decimal Tenor { get; set; }
+    public double Tenor { get; set; }
     public DateTime CreatedOn { get; set; }
     public decimal Yield { get; set; }
     public decimal Variance { get; set; }
