+++ Quotations Board Backend/DTOs/TBill/NewTbill.cs
@@ -1,6 +1,6 @@
 public class NewTbill
 {
-    public decimal Tenor { get; set; }
+    public double Tenor { get; set; }
     public DateTime IssueDate { get; set; }
     public decimal Yield { get; set; }
 }
