+++ Quotations Board Backend/Models/TBill/TBill.cs
@@ -9,7 +9,7 @@ public class TBill
     public string IssueNumber { get; set; } = null!;
     public DateTime IssueDate { get; set; }
     public DateTime MaturityDate { get; set; }
-    public decimal Tenor { get; set; }
+    public double Tenor { get; set; }
     public string CreatedBy { get; set; } = null!;
     public DateTime CreatedOn { get; set; }
     public decimal Yield { get; set; }
